Move boltdb compactor into subdirectory of boltdb package
