+++ pkg/storage/stores/shipper/boltdb/compactor/compacted_index.go
